requirements.txt
colorama
eth-brownie